Fix failure when using Alpha branch of Test::More
The Alpha branch of Test-Simple and friends is quickly approaching
stable status. It is already merged into blead perl to find issues like
this.

This patch makes your test suite pass with both old and new versions of
Test::Builder.

Please see https://github.com/Test-More/test-more/issues/474  for the
discussion of this issue.